+++ src/ui/activities/filetransfer_activity/misc.rs
@@ -130,11 +130,19 @@ impl FileTransferActivity {
     ///
     /// Build explorer reading configuration from `ConfigClient`
     pub(super) fn build_explorer(cli: Option<&ConfigClient>) -> FileExplorer {
-        FileExplorerBuilder::new()
-            .with_file_sorting(FileSorting::ByName)
-            .with_group_dirs(Some(GroupDirs::First))
-            .with_stack_size(16)
-            .build()
+        match &cli {
+            Some(cli) => FileExplorerBuilder::new() // Build according to current configuration
+                .with_file_sorting(FileSorting::ByName)
+                .with_group_dirs(cli.get_group_dirs())
+                .with_hidden_files(cli.get_show_hidden_files())
+                .with_stack_size(16)
+                .build(),
+            None => FileExplorerBuilder::new() // Build default
+                .with_file_sorting(FileSorting::ByName)
+                .with_group_dirs(Some(GroupDirs::First))
+                .with_stack_size(16)
+                .build(),
+        }
     }
 
     /// ### setup_text_editor